MAC Address Changer Tools
Unlike we think, IP address is not the only factor that helps connect computers in a network, but MAC address is also equally significant. So, whether you are connected using a wireless network or through a wired one, a network card is needed for your computer to be connected to others and vice a verse. The network is assigned a specific address or attribute that are usually allotted by the manufacturers and this called Media Access Control or MAC address. However, you can change them manually later with the MAC spoofing process.

Significance of MAC Address Changer Tools

While we already know that the MAC address changer tools help you to change the unique addresses of your network adapters, it primarily helps protect your network from needless connections. You simply need to approve the MAC addresses you want to link to the network or the internet and that’s it.

While you can certainly change the MAC address of the network adapter manually through the control panel, there are quite a few software that helps you to take care of it easily, without any mess and for free! So, if you are looking to change the MAC address for verification, privacy, limited data usage, etc. you can surely make use of any of the below listed best free MAC address changer tools for your purpose.


Follow the command:
==============

+ Download the mac_changer tool on GitHub link
+ chmod +x mac_changer.py
+ python --help (or) -h
How to run mac_changer:
--------------------------
+ python mac_changer.py -i <interface> -m <enter the new mac>